move copy function into utils
