+++ client/src/JoinGame/index.js
@@ -8,6 +8,7 @@ import {
   hasSufficientNumbersToPlay,
   isNameAvailable
 } from "../utils/gameUtils";
+import { copyText } from "../utils/utils";
 
 import "./styles.css";
 
@@ -35,37 +36,13 @@ class JoinGame extends Component {
   };
 
   copyInviteLink = () => {
-    const textarea = document.createElement("textarea");
-    textarea.textContent = window.location.href;
-    textarea.contentEditable = "true";
-    textarea.readOnly = false;
-    textarea.disabled = false;
-    // Can't be `display: none` for `select()` to work.
-    textarea.setAttribute(
-      "style",
-      "opacity: 0; position: fixed; top: -1000px; left: -1000px;"
-    );
-    // Node needs to actually be in the live DOM during copy...
-    if (document.body !== null) {
-      document.body.appendChild(textarea);
-    }
-    textarea.select();
-    let didCopy = document.execCommand("copy");
-    if (!didCopy) {
-      // iOS-specific copying
-      // Borrowed from http://bit.ly/2Gc3gTp
-      const range = document.createRange();
-      range.selectNodeContents(textarea);
-      const selection = window.getSelection();
-      selection.removeAllRanges();
-      selection.addRange(range);
-      textarea.setSelectionRange(0, 999999);
-      didCopy = document.execCommand("copy");
-    }
-    if (document.body !== null) {
-      document.body.removeChild(textarea);
-    }
+    copyText();
     this.registerCopySuccess();
+  };
+
+  registerCopySuccess = () => {
+    this.toggleCopyButtonText();
+    setTimeout(this.toggleCopyButtonText, 3000);
   };
 
   toggleCopyButtonText = () => {
@@ -75,11 +52,6 @@ class JoinGame extends Component {
         copyButtonText: copyButtonText === copied ? "Copy" : copied
       };
     });
-  };
-
-  registerCopySuccess = () => {
-    this.toggleCopyButtonText();
-    setTimeout(this.toggleCopyButtonText, 3000);
   };
 
   render() {
